requirements.txt
boto3==1.20.28
botocore==1.23.28
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
ConfigArgParse==1.5.3
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.3
greenlet==1.1.2
idna==3.3
invokust==0.76
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
locust==2.5.1
MarkupSafe==2.0.1
msgpack==1.0.3
numpy==1.22.0
pandas==1.3.5
Pillow==9.0.0
psutil==5.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.0
roundrobin==0.0.2
s3transfer==0.5.0
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
